Manage number of instances when editing products
There now exists a Product#quantity field that determines the number of
instances that must be alive after save. This is analoguous to the
classic ActiveRecord counter_cache, except that quantity must not be
readonly: we want to increase and decrease the value and reflect that
in the actual instance rows.

Care was taken to not remove instances that have had reservations made
against them. This keeps as much data as possible, so that full history
can be tracked.