Improve reservations page UX by not scrolling back on save
Instead of doing a full form submit and page redraw, only redraw the
singel TR that holds the row that was just modified. This improves the
user experience 100x because we can simply scroll down the list and add
and remove products/instnaces to our event without losing our position
on the view.