Notify people as events progress through their lifecycle
Events can be in one of four states:

* Draft
* Finalized
* Ready
* Audited

In order to reduce load on the inventory directors, the system
automatically sends emails to people depending on what state we are in.
For example, when entering the Audited state, accountants, troop members
and inventory directors automatically receive an email with the final
invoice.