Introduce concept of consumables
Consumables are "products" that are consumed after being taken out of
the inventory. Think propane, juice, cans of soup, raw minced meat,
garbage bags, etc. The available quantity of a given consumable is the
sum of it's transactions (a model that will be added shortly).

Consumables were designed to duck-type as products, in order to reuse as
much code as possible.