Display consumables alongside products on event/reservation
At the moment, consumables are hard-coded to show up first, but we
should eventually mix consumables and products. Filtering by string
works for both products and consumables. We cannot record usage of
consumables, but that is coming later.

This commit also introduces a completely redesigned event/reservations
page. This page is closer to a shop, and pictures are front and center
for all products. This makes it much easier to browse the catalog and
see what's available.